requirements.txt
opencv-python
mediapipe
numpy
sounddevice
SpeechRecognition
PyAudio
pyttsx3
PyYAML
mysql-connector-python
bcrypt
Flask
